Rename utils file and modify test
